+++ auth_manager.go
@@ -27,6 +27,7 @@ type SessionHandler interface {
 	GetSessionDetail(keyName string) (SessionState, bool)
 	GetSessions(filter string) []string
     GetStore() StorageHandler
+    ResetQuota(string, SessionState)
 }
 
 type KeyGenerator interface {
@@ -86,6 +87,12 @@ func (b *DefaultSessionManager) Init(store StorageHandler) {
 
 func (b *DefaultSessionManager) GetStore() StorageHandler{
 	return b.Store
+}
+
+func (b *DefaultSessionManager) ResetQuota(keyName string, session SessionState) {
+    log.Warning("Tracked quota reset for key: ", keyName)
+	rawKey := QuotaKeyPrefix + keyName
+    b.Store.SetKey(rawKey, "0", session.QuotaRenewalRate)
 }
 
 // UpdateSession updates the session state in the storage engine